refactor:refraction of the company entity
