+++ examples/mock_example.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/brazzcore/mocai/pkg/mocai/constants"
 	"github.com/brazzcore/mocai/pkg/mocai/entities/address"
-	company "github.com/brazzcore/mocai/pkg/mocai/entities/company/brazil"
+	"github.com/brazzcore/mocai/pkg/mocai/entities/company"
 	"github.com/brazzcore/mocai/pkg/mocai/entities/person"
 	"github.com/brazzcore/mocai/pkg/mocai/entities/phone"
 	"github.com/brazzcore/mocai/pkg/mocai/translations"
@@ -43,7 +43,7 @@ func GenerateMockExample() {
 	fmt.Printf("Person: %s %s, %s, %d years old, CPF: %s\n",
 		person_mock.FirstNameMale, person_mock.LastName, person_mock.Gender, person_mock.Age, person_mock.CPF)
 
-	fmt.Printf("Company: %s, CNPJ: %s\n", company_mock.CompanyName, company_mock.CNPJ)
+	fmt.Printf("Company: %s, CNPJ: %s\n", company_mock.BrazilianCompany.CompanyName, company_mock.BrazilianCompany.CNPJ)
 
 	fmt.Printf("Address: %s, %d - %s, %s (%s) - %s\n",
 		address_mock.Street, address_mock.Number, address_mock.City, address_mock.State, address_mock.UF, address_mock.ZIP)
